Refactoring on instances API
